+++ supabase/functions/_utils/stripe.ts
@@ -1,8 +1,9 @@
 import axios from 'https://deno.land/x/axiod@0.26.2/mod.ts'
+import { getEnv } from './utils.ts'
 
 const getAuth = () => {
   // get stripe token
-  const STRIPE_SECRET_KEY = Deno.env.get('STRIPE_SECRET_KEY') || ''
+  const STRIPE_SECRET_KEY = getEnv('STRIPE_SECRET_KEY') || ''
   const STRIPE_TOKEN = `${STRIPE_SECRET_KEY}:`
   // encode b64
   const STRIPE_TOKEN_B64 = btoa(STRIPE_TOKEN)
@@ -69,7 +70,7 @@ export const createCustomer = async (email: string) => {
 
 export const removeOldSubscription = async (subscriptionId: string) => {
   console.log('removeOldSubscription', subscriptionId)
-  const STRIPE_SECRET_KEY = Deno.env.get('STRIPE_SECRET_KEY') || ''
+  const STRIPE_SECRET_KEY = getEnv('STRIPE_SECRET_KEY') || ''
   const STRIPE_TOKEN = `${STRIPE_SECRET_KEY}`
   console.log('STRIPE_TOKEN', STRIPE_TOKEN)
   const requestOptions = {