+++ supabase/functions/channel/index.ts
@@ -1,7 +1,8 @@
 import { serve } from 'https://deno.land/std@0.167.0/http/server.ts'
 import { checkAppOwner, supabaseAdmin, updateOrCreateChannel } from '../_utils/supabase.ts'
 import type { Database } from '../_utils/supabase.types.ts'
-import { checkKey, fetchLimit, sendRes } from '../_utils/utils.ts'
+import type { BaseHeaders } from '../_utils/types.ts'
+import { checkKey, fetchLimit, methodJson, sendRes } from '../_utils/utils.ts'
 
 interface ChannelSet {
   app_id: string
@@ -22,9 +23,7 @@ interface GetDevice {
   page?: number
 }
 
-export const get = async (event: Request, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> => {
-  const url = new URL(event.url)
-  const body = Object.fromEntries(url.searchParams.entries() as any) as GetDevice
+export const get = async (body: GetDevice, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> => {
   if (!body.app_id || !(await checkAppOwner(apikey.user_id, body.app_id)))
     return sendRes({ status: 'You can\'t access this app', app_id: body.app_id }, 400)
 
@@ -89,9 +88,7 @@ export const get = async (event: Request, apikey: Database['public']['Tables']['
   }
 }
 
-export const deleteChannel = async (event: Request, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> => {
-  const body = (await event.json()) as ChannelSet
-
+export const deleteChannel = async (body: ChannelSet, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> => {
   if (!(await checkAppOwner(apikey.user_id, body.app_id)))
     return sendRes({ status: 'You can\'t access this app', app_id: body.app_id }, 400)
 
@@ -110,9 +107,7 @@ export const deleteChannel = async (event: Request, apikey: Database['public']['
   return sendRes()
 }
 
-export const post = async (event: Request, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> => {
-  const body = (await event.json()) as ChannelSet
-
+export const post = async (body: ChannelSet, apikey: Database['public']['Tables']['apikeys']['Row']): Promise<Response> => {
   const channel: Database['public']['Tables']['channels']['Insert'] = {
     created_by: apikey.user_id,
     app_id: body.app_id,
@@ -152,8 +147,8 @@ export const post = async (event: Request, apikey: Database['public']['Tables'][
   return sendRes()
 }
 
-serve(async (event: Request) => {
-  const apikey_string = event.headers.get('authorization')
+const main = async (url: URL, headers: BaseHeaders, method: string, body: any) => {
+  const apikey_string = headers.authorization
 
   if (!apikey_string)
     return sendRes({ status: 'Missing apikey' }, 400)
@@ -163,16 +158,29 @@ serve(async (event: Request) => {
     if (!apikey)
       return sendRes({ status: 'Missing apikey' }, 400)
 
-    if (event.method === 'POST')
-      return post(event, apikey)
-    else if (event.method === 'GET')
-      return get(event, apikey)
-    else if (event.method === 'DELETE')
-      return deleteChannel(event, apikey)
+    if (method === 'POST')
+      return post(body, apikey)
+    else if (method === 'GET')
+      return get(body, apikey)
+    else if (method === 'DELETE')
+      return deleteChannel(body, apikey)
   }
   catch (e) {
     return sendRes({ status: 'Error', error: JSON.stringify(e) }, 500)
   }
 
   return sendRes({ status: 'Method now allowed' }, 400)
+}
+
+serve(async (event: Request) => {
+  try {
+    const url: URL = new URL(event.url)
+    const headers: BaseHeaders = Object.fromEntries(event.headers.entries())
+    const method: string = event.method
+    const body: any = methodJson.includes(method) ? await event.json() : Object.fromEntries(url.searchParams.entries())
+    return main(url, headers, method, body)
+  }
+  catch (e) {
+    return sendRes({ status: 'Error', error: JSON.stringify(e) }, 500)
+  }
 })