+++ supabase/functions/cron_good_plan/index.ts
@@ -1,9 +1,9 @@
 import { serve } from 'https://deno.land/std@0.167.0/http/server.ts'
 import { supabaseAdmin } from '../_utils/supabase.ts'
-import { sendRes } from '../_utils/utils.ts'
+import { getEnv, sendRes } from '../_utils/utils.ts'
 
 serve(async (event: Request) => {
-  const API_SECRET = Deno.env.get('API_SECRET')
+  const API_SECRET = getEnv('API_SECRET')
   const authorizationSecret = event.headers.get('apisecret')
   if (!authorizationSecret)
     return sendRes({ status: 'Cannot find authorization secret' }, 400)