+++ supabase/functions/on_channel_create/index.ts
@@ -2,10 +2,10 @@ import { serve } from 'https://deno.land/std@0.167.0/http/server.ts'
 import type { InsertPayload } from '../_utils/supabase.ts'
 import { supabaseAdmin, updateOrAppStats } from '../_utils/supabase.ts'
 import type { Database } from '../_utils/supabase.types.ts'
-import { sendRes } from '../_utils/utils.ts'
+import { getEnv, sendRes } from '../_utils/utils.ts'
 
 serve(async (event: Request) => {
-  const API_SECRET = Deno.env.get('API_SECRET')
+  const API_SECRET = getEnv('API_SECRET')
   const authorizationSecret = event.headers.get('apisecret')
   if (!authorizationSecret || !API_SECRET || authorizationSecret !== API_SECRET)
     return sendRes({ message: 'Fail Authorization' }, 400)