+++ supabase/functions/stripe_event/index.ts
@@ -2,12 +2,12 @@ import { serve } from 'https://deno.land/std@0.167.0/http/server.ts'
 import { addDataPerson, addEventPerson, updatePerson } from '../_utils/crisp.ts'
 import { extractDataEvent, parseStripeEvent } from '../_utils/stripe_event.ts'
 import { supabaseAdmin } from '../_utils/supabase.ts'
-import { sendRes } from '../_utils/utils.ts'
+import { getEnv, sendRes } from '../_utils/utils.ts'
 import { removeOldSubscription } from '../_utils/stripe.ts'
-import { logsnag } from '../_utils/_logsnag.ts'
+import { logsnag } from '../_utils/logsnag.ts'
 
 serve(async (event: Request) => {
-  if (!event.headers.get('stripe-signature') || !Deno.env.get('STRIPE_WEBHOOK_SECRET') || !Deno.env.get('STRIPE_SECRET_KEY'))
+  if (!event.headers.get('stripe-signature') || !getEnv('STRIPE_WEBHOOK_SECRET') || !getEnv('STRIPE_SECRET_KEY'))
     return sendRes({ status: 'Webhook Error: no signature or no secret found' }, 400)
 
   // event.headers